refactor/#245: 반려동물 정보 등록 진행상황 관련 코드 훅으로 분리
